+++ app/checks/functionality/task.js
@@ -29,7 +29,7 @@
         __MarkbotInjectedFunctions.failLabel = '__markbot-functionality-test-fail-${label}';
         __MarkbotInjectedFunctions.debugLabel = '__markbot-functionality-test-debug-${label}';
 
-        (function ($, $$, css, on, ev, send, hover, pass, fail, debug) {
+        (function ($, $$, css, bounds, on, ev, send, hover, pass, fail, debug) {
           'use strict';
 
           try {
@@ -42,6 +42,7 @@
           __MarkbotInjectedFunctions.$,
           __MarkbotInjectedFunctions.$$,
           __MarkbotInjectedFunctions.css,
+          __MarkbotInjectedFunctions.bounds,
           __MarkbotInjectedFunctions.on,
           __MarkbotInjectedFunctions.ev,
           __MarkbotInjectedFunctions.send,